==> pilot/code/05_create_survey.ipynb <==
import pandas as pd

# Load responses
df = pd.read_csv('completions/all_responses.csv')

with open('pilot_completions.txt', 'w') as f:
    f.write('[[AdvancedFormat]]\n\n')
    
    for index, row in df.iterrows():
        response = row['response']
        response_id = row['response_id']
        model = row['model']
        issue_short = row['issue_short']
        prompt_variant_number = row['prompt_variant_number']
        
        # Replace line breaks with <br> for Qualtrics
        response = response.replace('\n', '<br>')
        
        # Write to file in the required format
        f.write(f'[[Block:{model}-{issue_short}-{prompt_variant_number}]]\n')
        f.write('[[Question:Text]]\n')
        f.write(f'{response}\n')
        f.write(f'[[ED:response_id:{response_id}]]\n\n')
